moving wordcount functions to separate file so people can focus on the command line args
